requirements.txt
discord.py==2.6.0
requests==2.31.0
python-dotenv==1.1.1